refactor: modify Class name Result to Output
